Removes IXC as dispatcher from examples
This patch removes `IXC` as RTIC task dispatcher from the examples.

`IXC` is described in the user manual as "Floating Point Unit Interrupt IXC
associated with FPU cumulative exception bit." and may at first glance seem as
a niche interrupt that generally won't trigger. However, code as that used in
the `apply_duty` function of the PWM module will--in certain cases--trigger
this exception and then RTIC tasks using the related dispatcher will cease
scheduling silently.

For small programs that only demonstrate single peripherals, the likelihood
of actually triggering this interrupt is minimal and is not really a problem.
On the other hand, the users who look at these examples may not understand
what `IXC` means and decide to use it for their applications because it's been
chosen as a default by others.

Then, when they initialize some PWM channels and the setup routines trigger
the interrupt, the scheduler will start misbehaving with only (seemingly)
innocuous floating-point operations to blame.